Joinpoint Regression Models by Groups

Fits segmented linear regression models by groups for age-standardized rates using joinpoint regression. Models can be fitted using either a stepwise selection procedure based on the Bayesian Information Criterion (BIC) or a fixed number of joinpoints. Internally calls segmented::selgmented() or segmented::segmented() and applies a log transformation to the response variable.

model_jp(data, value, time, group, k = 2, step = TRUE, test = TRUE)

Arguments

data

Data frame containing age-standardized rates.

value

Name of the response variable (character).

time

Name of the time variable (character).

group

Name of one or more grouping variables (character vector).

k

Maximum number of joinpoints to be estimated (integer).

step

Use stepwise procedure to select the number of joinpoints (logical).

test

whether to test for differences in slope t-values during the stepwise selection procedure. Only used when step = TRUE.

Value

A list of models by group.

Author

Tamara Ricardo

Examples

# Generate example data
library(dplyr)
df <- mtcars |>
mutate(
  year = seq(2000, length.out = n(), by = 1),
  group = factor(paste("cyl", cyl, sep = "_")),
  rate = mpg
) |>
select(year, group, rate)

# Check group levels
levels(df$group)
#> [1] "cyl_4" "cyl_6" "cyl_8"

# Fit models
mods <- model_jp(data = df, value = "rate", time = "year", group = "group",
 k = 2, step = TRUE, test = TRUE)
#> Model: cyl_6 | Joinpoint(s): no significant joinpoints detected
#> Model: cyl_4 | Joinpoint(s): no significant joinpoints detected
#> Model: cyl_8 | Joinpoint(s): no significant joinpoints detected

# Show the output of the first model
mods$cyl_6
#> 
#> Call:
#> stats::lm(formula = formula, data = .x)
#> 
#> Coefficients:
#> (Intercept)         year  
#>   -3.820829     0.003514  
#> 
summary(mods$cyl_6)
#> 
#> Call:
#> stats::lm(formula = formula, data = .x)
#> 
#> Residuals:
#>      Min       1Q   Median       3Q      Max 
#> -0.25210 -0.09738 -0.03658  0.12882  0.25009 
#> 
#> Coefficients:
#>              Estimate Std. Error t value Pr(>|t|)
#> (Intercept) -3.820829  12.032563  -0.318    0.758
#> year         0.003514   0.005962   0.589    0.570
#> 
#> Residual standard error: 0.1734 on 9 degrees of freedom
#> Multiple R-squared:  0.03716,	Adjusted R-squared:  -0.06982 
#> F-statistic: 0.3473 on 1 and 9 DF,  p-value: 0.5701
#>