Joinpoint Regression Models by Groups

Fits segmented linear regression models by groups for age-standardized rates using joinpoint regression. Models can be fitted using either a stepwise selection procedure based on the Bayesian Information Criterion (BIC) or a fixed number of joinpoints. Internally, the function calls segmented::selgmented() or segmented::segmented() and applies a log transformation to the response variable.

model_jp(data, value, time, group, k = 2, step = TRUE, test = TRUE)

Arguments

data

A data frame containing age-standardized rates.

value

Response variable.

time

Time variable.

group

Names of one or more grouping variables.

k

Maximum number of joinpoints to estimate.

step

Logical. If TRUE, uses a stepwise procedure to select the number of joinpoints based on BIC. If FALSE, fits a model with a fixed number of joinpoints specified by k.

test

Logical. If TRUE, tests differences in slope t-values during the stepwise selection procedure. Only used when step = TRUE.

Value

A named list of joinpoint regression models by group.

Details

The National Cancer Institute (NCI) recommends the following maximum number of joinpoints according to the length of the time series (Kim et al., 2000):

0–6 time points: 0 joinpoints.

7–11 time points: 1 joinpoint.

12–16 time points: 2 joinpoints.

17–21 time points: 3 joinpoints.

22–26 time points: 4 joinpoints.

27–31 time points: 5 joinpoints.

32–36 time points: 6 joinpoints.

37 or more time points: 7 joinpoints.

#' @references Kim HJ, Fay MP, Feuer EJ, Midthune DN (2000). "Permutation Tests for Joinpoint Regression with Applications to Cancer Rates." Statistics in Medicine, 19(3), 335–351. doi:10.1002/(sici)1097-0258(20000215)19:3<335::aid-sim336>3.0.co;2-z.

Author

Tamara Ricardo

Examples

# Load example data
data("hiv_data")

# Check group levels
levels(hiv_data$region)
#> [1] "Central" "East"    "North"   "South"   "West"   

# Fit models
mods <- model_jp(data = hiv_data, value = hiv_rate, time = year, group = c("region", "sex"),
 k = 2, step = TRUE, test = TRUE)
#> Model: Central_Female | Joinpoint(s): 2019
#> Model: Central_Male | Joinpoint(s): 2013.7; 2019.0
#> Model: East_Female | Joinpoint(s): 2014.8; 2019.1
#> Model: East_Male | Joinpoint(s): 2014.4; 2019.0
#> Model: North_Female | Joinpoint(s): 2015.4; 2019.0
#> Model: North_Male | Joinpoint(s): 2014.9; 2019.1
#> Model: South_Female | Joinpoint(s): 2019
#> Model: South_Male | Joinpoint(s): 2019
#> Model: West_Female | Joinpoint(s): 2017.9; 2019.4
#> Model: West_Male | Joinpoint(s): 2019

# Show the output of the first model by calling its index
mods[[1]]
#> Call: segmented::selgmented(olm = stats::lm(formula = log(hiv_rate) ~ 
#>     year), Kmax = 2)
#> 
#> Coefficients of the linear terms:
#> (Intercept)     .jp_time  U1..jp_time  
#>   102.04734     -0.04836      0.08174  
#> 
#> Estimated Break-Point(s):
#> psi1..jp_time  
#>          2019  

# Same output will be obtained when calling model name
mods$Central
#> NULL